+++ examples/aws_ha_fortigate/main.tf
@@ -1,6 +1,6 @@
 module "transit_ha_aws" {
   source  = "terraform-aviatrix-modules/mc-transit/aviatrix"
-  version = "v2.3.0"
+  version = "v2.3.4"
 
   cloud                  = "aws"
   name                   = "transit-ha-aws"