+++ examples/gcp_ha_palo_alto/main.tf
@@ -1,6 +1,6 @@
 module "transit_ha_gcp" {
   source  = "terraform-aviatrix-modules/mc-transit/aviatrix"
-  version = "2.3.0"
+  version = "2.3.4"
 
   cloud                  = "gcp"
   name                   = "transit-ha-gcp"
@@ -19,5 +19,5 @@ module "mc_firenet_ha_gcp" {
   firewall_image = "Palo Alto Networks VM-Series Next-Generation Firewall BUNDLE1"
   egress_enabled = true
   egress_cidr    = "10.102.1.0/24"
-  mgmt_cidr      = "10.102.3.0/24"
+  mgmt_cidr      = "10.102.3.4/24"
 }